+++ docs/index.html
@@ -1,3 +1,3 @@
 <head>
-    <meta content="0; url=https://eliza-diggins.github.io/cluster_generator/build/html" http-equiv="refresh"/>
+    <meta content="0; url=https://jzuhone.github.io/cluster_generator/build/html" http-equiv="refresh"/>
 </head>
