+++ docs/source/correction.rst
@@ -3,7 +3,7 @@
 Correcting Non-Physical Behavior
 ================================
 
-In many instances, non-physical behaviors can arise in the :py:class:`model.ClusterModel` instances generated by the CGP. This
+In many instances, non-physical behaviors can arise in the :py:class:`model.ClusterModel` instances generated by CG. This
 is usually because of some physical consideration that was not taken into account by the modeler; however, it is often difficult (as the
 modeler) to identify the issues with your model in order to correct them. On this page, we will walk you through the different non-physicalities that
 commonly occur and how to fix them.
@@ -11,7 +11,7 @@ commonly occur and how to fix them.
 What Causes Non-Physical Behavior?
 ----------------------------------
 
-The first thing that all CGP users should have a good understanding of is how different non-physicality arise in our models and
+The first thing that all CG users should have a good understanding of is how different non-physicality arise in our models and
 what each really means from a physical level. There are typically 2 sources of non-physicality (Type 1 and Type 2). In Type 1, the underlying Newtonian model
 is not self consistent because of some feature of the profiles used to generate the model. These are the most common type of non-physicality. Type 2 non-physicalities are
 an exclusive feature of MONDian clusters generated from profiles which would otherwise be suitable for use in Newtonian dynamics. We will described each of these types of
@@ -108,7 +108,7 @@ used to produce the model.
 Type 2 Non-Physicalities
 ''''''''''''''''''''''''
 
-Type 2 Non-physicalities are gravity theory specific. Due to the modularity of the gravity implementation and the capacity for the CGP to support arbitrary
+Type 2 Non-physicalities are gravity theory specific. Due to the modularity of the gravity implementation and the capacity for CG to support arbitrary
 gravitational theories, we will not exhaustively cover niche cases here. Documentation for each of the individual gravity theories should include a table with
 the relevant possible non-physicalities.
 
@@ -133,7 +133,7 @@ We note the existence of these non-physicalities because they will, inevitably, 
 .. figure:: _images/diagrams/non-physical.drawio.png
     :alt: Non-physicalities diagram.
 
-    All of the possible non-physicalities which occur in the CGP.
+    All of the possible non-physicalities which occur in CG.
 
 Fixing Non-Physical Behaviors
 -----------------------------
