+++ docs/source/index.rst
@@ -7,20 +7,20 @@ cluster_generator
 
    <hr style="height:10px;background-color:black">
 
-The `Cluster Generator Project <https://eliza-diggins.github.io/cluster_generator>`_ (CGP) is a cross-platform Python library for generating initial conditions of galaxy clusters for N-body / hydrodynamics codes.
-CGP provides a variety of construction approaches, different physical assumption, profiles, and gravitational theories. Furthermore, the CGP is intented to interface with
-the vast majority of N-body / hydrodynamics codes, reducing the headache of converting initial conditions between formats for different simulation softwares. GCP's goal is to provide
+`Cluster Generator Project <https://jzuhone.github.io/cluster_generator>`_ (CG) is a cross-platform Python library for generating initial conditions of galaxy clusters for N-body / hydrodynamics codes.
+CG provides a variety of construction approaches, different physical assumption, profiles, and gravitational theories. Furthermore, CG is intended to interface with
+a number of N-body / hydrodynamics codes used in studies of galaxy clusters, reducing the headache of converting initial conditions between formats for different simulation softwares. GCP's goal is to provide
 comprehensive tools for modeling and implementation of galaxy clusters in astrophysical simulations to promote the study of galaxy cluster dynamics.
 
 This repository contains the core package, which is constructed modularly to facilitate easy development by users to meet particular scientific use cases. All of the
 necessary tools to get started building initial conditions are provided.
 
-You can access the documentation `here <http://eliza-diggins.github.io/cluster_generator>`_, or build it from scratch using the ``./docs`` directory in this source distribution.
+You can access the documentation `here <http://jzuhone.github.io/cluster_generator>`_, or build it from scratch using the ``./docs`` directory in this source distribution.
 
-Developement occurs here on Github, if you encounter any bugs, issues, documentation failures, or want to suggest features, we recommend that you submit an issue on
+Development occurs here on Github, if you encounter any bugs, issues, documentation failures, or want to suggest features, we recommend that you submit an issue on
 the issues page of the repository.
 
-For installation directions, visit the `getting started page <https://eliza-diggins.github.io/cluster_generator/build/html/Getting_Started.html>`_.
+For installation directions, visit the `getting started page <https://jzuhone.github.io/cluster_generator/build/html/Getting_Started.html>`_.
 
 
 .. raw:: html
@@ -85,7 +85,7 @@ Features
 
         .. dropdown:: Gravitational Theories
 
-            ``cluster_generator`` provides not only high quality initial condition generation capacity, but also provides a
+            ``cluster_generator`` not only provides initial condition generation capacity, but also provides a
             comprehensive catalog of alternative gravity theories to explore. The following are built-in, but adding more
             is a relatively simple task:
 
@@ -249,7 +249,7 @@ Related Projects
         ^^^^
 
         Coupled with PyXSIM, SOXS is a instrument simulation tool for turning mock photon counts into realistic X-ray observations specific to the
-        behavior of specific instruments like CHANDRA, XMM-Newton, and NuSTAR.
+        behavior of specific instruments like *CHANDRA*, *XMM-Newton*, and *NuSTAR*.
 
         +++
 
@@ -280,7 +280,7 @@ Indices and tables
    :target: https://yt-project.org
 
 .. |docs| image:: https://img.shields.io/badge/docs-latest-brightgreen.svg
-   :target: https://eliza-diggins.github.io/cluster_generator/build/html/index.html
+   :target: https://jzuhone.github.io/cluster_generator/build/html/index.html
 
 .. |testing| image:: https://github.com/Eliza-Diggins/cluster_generator/actions/workflows/test.yml/badge.svg
 .. |Pylint| image:: https://github.com/Eliza-Diggins/cluster_generator/actions/workflows/pylint.yml/badge.svg
@@ -288,4 +288,4 @@ Indices and tables
 .. |coverage| image:: https://coveralls.io/repos/github/Eliza-Diggins/cluster_generator/badge.svg
    :target: https://coveralls.io/github/Eliza-Diggins/cluster_generator
 .. |ncodes| image:: https://img.shields.io/static/v1?label="Implemented%20Sim.%20Codes"&message="7"&color="red"
-    :target: https://eliza-diggins.github.io/cluster_generator/build/html/codes.html
+    :target: https://jzuhone.github.io/cluster_generator/build/html/codes.html
